# Faster JSON serialising of labels