@bboreham 2023-07-24
Upstream version: https://github.com/prometheus/prometheus/pull/12598
(Mimir should prefer that, if it gets merged)